+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 0.0.7
+- Add CampfireSafeZoneRadius setting.
+
 ## 0.0.6
 - Add opt-out setting for greatly improving performance of dynamite explosions.